aplicando arquitectura limpia
