--- copilot.txt
@@ -1,5 +1,0 @@
-
-ME: Que hora es?
-AI: 3:15pm
-ME: Xd
-AI: 3:15pm
