+++ routes/__init__.py
@@ -1,0 +1,6 @@
+from routes.auth_routes import auths
+from routes.download_routes import downloads
+from routes.examples_routes import examples
+from routes.project_routes import project
+from routes.static_projects_routes import statics
+from routes.user_routes import user
