+++ services/documentation/__init__.py
@@ -1,0 +1,1 @@
+from services.documentation.find_document import find_document
