+++ services/documentation/find_document.py
@@ -1,0 +1,8 @@
+from libs import Mongodb
+from bson.objectid import ObjectId
+
+def find_document(query):
+    mongo = Mongodb().client
+
+    user = mongo.db.documentation.find_one(query)
+    return user
