+++ services/project/find_project.py
@@ -1,0 +1,9 @@
+from libs import Mongodb
+from bson.objectid import ObjectId
+
+def find_project(query):
+    mongo = Mongodb().client
+
+    user = mongo.db.projects.find_one(query)
+    
+    return user
