+++ services/user/find_users.py
@@ -1,0 +1,7 @@
+from libs import Mongodb
+
+def find_users(query):
+    mongo = Mongodb().client
+
+    user = mongo.db.users.find(query)
+    return user
