+++ static/styles/profile.css
@@ -81,13 +81,14 @@ h1,h2,h3,h4,h5,p{
 
 @media screen and (min-width: 996px) {
     .custom-container {
-        margin-left: 40px;
+        margin: 0 auto;
+        width: 80%;
     }
 
-    .hero_profile {
+    /* .hero_profile {
         display: flex;
         justify-content: space-evenly;
-    }
+    } */
     
     .row {
         margin-top: 15px;
