+++ templates/user/profile/index.html
@@ -54,7 +54,7 @@
             <h3>Agrega un proyecto</h3>
             <a href="/add-project"  class="btn btn-primary">Agregar proyecto</a>
 
-            <div class='row' id="projects">
+            <div class="my-4" id="projects">
                 {% if user['projects'] %}
                 <div class="col col-lg-12">
                     <h3>Proyectos: {{ '0' if user['projects'] == None else user['projects'] | length }}</h3>
@@ -68,17 +68,17 @@
                             <p class='card-text'>
                                 {{project.description}}
                             </p>
-                            <div class="row" id="project-main-container">
-                                <div class="col">
+                            <div class="d-flex" id="project-main-container">
+                                <div class="col-4">
                                     <a href='/project/{{user.username}}/{{project.project_name}}' class='btn bi bi-pencil-fill'></a>
                                 </div>
-                                <div class="col">
+                                <div class="col-4">
                                     <form id="delete-project">
                                         <input type="hidden" name="id" value="{{project._id}}">
                                         <button class="btn bi bi-trash-fill btn-delete"></button>
                                     </form>
                                 </div>
-                                <div class="col">
+                                <div class="col-4">
                                     <a class='btn bi bi-file-arrow-down-fill' href="/download-project/{{project._id}}" download></a>
                                 </div>
                             </div>
