+++ utils/__init__.py
@@ -1,0 +1,2 @@
+from utils.cookies import destroyCookie
+from utils.times import timestamp
